CRUD

+++ src/main/java/com/dh/clinica/service/OdontologoServicio.java
@@ -1,5 +1,7 @@
 package com.dh.clinica.service;
 
+import com.dh.clinica.exceptions.BadRequestException;
+import com.dh.clinica.exceptions.ResourceNotFoundException;
 import com.dh.clinica.model.Odontologo;
 import com.dh.clinica.repository.impl.OdontologoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,24 +20,33 @@ public class OdontologoServicio {
         this.odontologoRepository = odontologoRepository;
     }
 
-    public Odontologo registrarOdontologo(Odontologo odontologo) {
+    public Odontologo registrar(Odontologo odontologo) throws BadRequestException{
+        if(odontologo == null)
+            throw  new BadRequestException("No pueden haber campos vacíos");
         return odontologoRepository.save(odontologo);
 
     }
 
-    public void eliminar(Integer id) {
+    public void eliminar(Integer id) throws ResourceNotFoundException, BadRequestException {
+        if(buscarPorId(id).isEmpty())
+            throw  new ResourceNotFoundException("No existe odontólogo con id: "+ id);
         odontologoRepository.deleteById(id);
     }
 
-    public Optional<Odontologo> buscar(Integer id) {
+    public Optional<Odontologo> buscarPorId(Integer id) throws ResourceNotFoundException, BadRequestException{
         return odontologoRepository.findById(id);
     }
 
-    public List<Odontologo> buscarTodos() {
+    public List<Odontologo> buscarTodos(){
         return odontologoRepository.findAll();
     }
 
-    public Odontologo actualizar(Odontologo odontologo) {
+    public Odontologo actualizar(Odontologo odontologo)throws BadRequestException {
+        if (odontologo == null)
+            throw new BadRequestException("No pueden haber campos vacíos");
+        if (odontologo.getId() == null)
+            throw new BadRequestException("El id del odontólogo no puede estar vacío");
+
         return odontologoRepository.save(odontologo);
     }
 }
